+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changes
 * [UI]: Fixed bug where SVG files could not be exported through the advanced visualization page. (0.21.2)
 * [UI]: Fixed bug where users could not share more than nine samples. (0.21.2)
 * [UI]: Moved the position of the notification system to top center.
+* [Workflow]: Added version 2.0.0 of a pipeline for running [bio_hansel](https://github.com/phac-nml/bio_hansel) (version 2.0.0)
 * [Workflow]: Added version 0.3 of a pipeline for running [SISTR](https://github.com/peterk87/sistr_cmd/) which now makes use of [Shovill](https://github.com/tseemann/shovill) for genome assembly.
 * [UI]: Users can save analysis results to samples after pipeline is done in "Share Results" tab.
 * [UI]: Fixed bug where edit groups page would throw a server exception. (0.21.3)